requirements.txt
absl-py==1.0.0
astunparse==1.6.3
attrs==21.4.0
autolab-core==1.1.0
cachetools==5.0.0
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.10
colorama==0.4.4
colorlog==6.6.0
cycler==0.11.0
dill==0.3.4
flatbuffers==2.0
fonttools==4.29.0
freetype-py==2.2.0
gast==0.4.0
google-auth==2.4.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.43.0
h5py==3.6.0
idna==3.3
imageio==2.14.1
imbalanced-learn==0.9.0
imblearn==0.0
importlib-metadata==4.10.1
joblib==1.1.0
jsonschema==4.4.0
keras==2.7.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
libclang==12.0.0
lxml==4.7.1
Markdown==3.3.6
matplotlib==3.5.1
mpmath==1.2.1
msgpack==1.0.3
multiprocess==0.70.12.2
networkx==2.6.3
numpy==1.22.1
oauthlib==3.1.1
opencv-python==4.5.5.62
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.0
Pillow==9.0.0
protobuf==3.19.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycollada==0.7.2
pyglet==1.5.21
PyOpenGL==3.1.0
pyparsing==3.0.7
pyreadline==2.1
pyrender==0.1.45
pyrsistent==0.18.1
python-dateutil==2.8.2
pytz==2021.3
PyWavelets==1.2.0
PyYAML==6.0
requests==2.27.1
requests-oauthlib==1.3.0
rsa==4.8
Rtree==0.9.7
ruamel.yaml==0.17.20
ruamel.yaml.clib==0.2.6
scikit-image==0.19.1
scikit-learn==1.0.2
scipy==1.7.3
seaborn==0.11.2
setproctitle==1.2.2
Shapely==1.8.0
six==1.16.0
sklearn==0.0
svg.path==4.1
sympy==1.9
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.7.0
tensorflow-estimator==2.7.0
tensorflow-io-gcs-filesystem==0.23.1
termcolor==1.1.0
threadpoolctl==3.0.0
tifffile==2021.11.2
trimesh==3.9.42
typing_extensions==4.0.1
urllib3==1.26.8
visualization==1.0.0
Werkzeug==2.0.2
wrapt==1.13.3
xxhash==2.0.2
zipp==3.7.0
